#5 Rename trading212.py to trading212_builder.py, avoids naming collision with API folder
